requirements.txt
discord.py==2.3.2
pymongo==4.6.0
requests==2.31.0
gradio==4.2.0
huggingface_hub==0.19.4
